With SmartSVN!! Smart (at least, by now, if you bought a license use it, it is so good to use supported applications for doing wellness)!! from MacOS VMware Workstation guest at Windows 8 at inha at veriverse at Curitiba - PR - Brazil!! Thank you God!! Thank you Carlos Gustavo Cecyn Lundgren!!
Obrigado Mummi!!

Obrigado Ralf Sieves!!

Obrigado Rafael Cecyn Lundgren!!

+++ appseed/base/base/openssl/crypto/asn1/a_sign.c
@@ -317,7 +317,7 @@ int ASN1_item_sign_ctx(const ASN1_ITEM *it,
 	if (signature->data != NULL) OPENSSL_free(signature->data);
 	signature->data=buf_out;
 	buf_out=NULL;
-	signature->length=outl;
+	signature->length=(int)outl;
 	/* In the interests of compatibility, I'll make sure that
 	 * the bit string has a 'not-used bits' value of 0
 	 */
@@ -329,5 +329,5 @@ err:
 		{ OPENSSL_cleanse((char *)buf_in,(unsigned int)inl); OPENSSL_free(buf_in); }
 	if (buf_out != NULL)
 		{ OPENSSL_cleanse((char *)buf_out,outll); OPENSSL_free(buf_out); }
-	return(outl);
+	return (int)(outl);
 	}
