+++ appseed/base/base/openssl/crypto/conf/conf_api.c
@@ -277,7 +277,7 @@ CONF_VALUE *_CONF_new_section(CONF *conf, const char *section)
 		goto err;
 	if ((v=OPENSSL_malloc(sizeof(CONF_VALUE))) == NULL)
 		goto err;
-	i=strlen(section)+1;
+	i=(int) (strlen(section)+1);
 	if ((v->section=OPENSSL_malloc(i)) == NULL)
 		goto err;
