+++ appseed/base/base/openssl/crypto/conf/conf_mod.c
@@ -328,9 +328,9 @@ static CONF_MODULE *module_find(char *name)
 	p = strrchr(name, '.');
 
 	if (p)
-		nchar = p - name;
+		nchar = (int) (p - name);
 	else 
-		nchar = strlen(name);
+		nchar = (int) strlen(name);
 
 	for (i = 0; i < sk_CONF_MODULE_num(supported_modules); i++)
 		{
@@ -552,7 +552,7 @@ char *CONF_get1_default_config_file(void)
 	if (file) 
 		return BUF_strdup(file);
 
-	len = strlen(X509_get_default_cert_area());
+	len = (int) strlen(X509_get_default_cert_area());
 #ifndef OPENSSL_SYS_VMS
 	len++;
 #endif
@@ -611,7 +611,7 @@ int CONF_parse_list(const char *list_, int sep, int nospc,
 				while(isspace((unsigned char)*tmpend))
 					tmpend--;
 				}
-			ret = list_cb(lstart, tmpend - lstart + 1, arg);
+			ret = list_cb(lstart, (int) (tmpend - lstart + 1), arg);
 			}
 		if (ret <= 0)
 			return ret;
