+++ appseed/base/base/openssl/ssl/t1_lib.c
@@ -428,7 +428,7 @@ unsigned char *ssl_add_clienthello_tlsext(SSL *s, unsigned char *p, unsigned cha
 	if (s->srp_ctx.login != NULL)
 		{ /* Add TLS extension SRP username to the Client Hello message */
 
-		int login_len = strlen(s->srp_ctx.login);	
+		int login_len = (int) strlen(s->srp_ctx.login);
 		if (login_len > 255 || login_len == 0)
 			{
 			SSLerr(SSL_F_SSL_ADD_CLIENTHELLO_TLSEXT, ERR_R_INTERNAL_ERROR);
@@ -504,7 +504,7 @@ unsigned char *ssl_add_clienthello_tlsext(SSL *s, unsigned char *p, unsigned cha
 		{
 		int ticklen;
 		if (!s->new_session && s->session && s->session->tlsext_tick)
-			ticklen = s->session->tlsext_ticklen;
+			ticklen = (int) s->session->tlsext_ticklen;
 		else if (s->session && s->tlsext_session_ticket &&
 			 s->tlsext_session_ticket->data)
 			{
@@ -662,7 +662,7 @@ unsigned char *ssl_add_clienthello_tlsext(SSL *s, unsigned char *p, unsigned cha
                 }
 #endif
 
-	if ((extdatalen = ret-p-2)== 0) 
+	if ((extdatalen = (int) (ret-p-2))== 0)
 		return p;
 
 	s2n(extdatalen,p);
@@ -851,7 +851,7 @@ unsigned char *ssl_add_serverhello_tlsext(SSL *s, unsigned char *p, unsigned cha
 		}
 #endif
 
-	if ((extdatalen = ret-p-2)== 0) 
+	if ((extdatalen = (int) (ret-p-2))== 0)
 		return p;
 
 	s2n(extdatalen,p);
